% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hla.R
\name{clean_hla}
\alias{clean_hla}
\title{Correct the format of HLA alleles}
\usage{
clean_hla(allele)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A string or character vector of the same length as \code{allele}, with the
cleaned-up alleles (or the originals, if no cleaning was necessary).
}
\description{
\code{clean_hla()} takes in a string or character vector of HLA alleles, and
performs a number of cleaning steps to correct common issues with formatting.
}
\details{
\code{clean_hla()} performs the following operations, in this order:
\enumerate{
\item Removes leading or trailing whitespace
\item Adds a leading zero to fields if necessary (\code{A*1:1} --> \code{A*01:01})
\item Removes redundant "versions" of the allele (e.g. a broad when the split is
also specified). See \code{\link[=strip_redundant]{strip_redundant()}}
\item Removes punctuation and symbols that are not part of the notation
\item Adds an ":XX" suffix to molecular alleles with only 1 field (\code{A*01} -->
\code{A*01:XX})
\item Propagates loci and allele group fields in ambiguities (\code{A*01:01/02} -->
\code{A*01:01/A*01:02})
\item Converts v2 to v3 (\verb{A*01010102N} --> \verb{A*01:01:01:02N}). See
\code{\link[=convert_v2_to_v3]{convert_v2_to_v3()}}
\item Converts deleted/changed alleles (\code{C*03:99:01} --> \code{C*01:169:01}). See
\code{\link[=convert_deleted]{convert_deleted()}}
}
}
\examples{
clean_hla(" A'*1 ")
clean_hla("A25(10)")
clean_hla("C*03:01/02")
}
