% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hla.R
\name{convert_deleted}
\alias{convert_deleted}
\title{Lookup the new designation for deleted or changed HLA alleles}
\usage{
convert_deleted(allele)
}
\arguments{
\item{allele}{A string or character vector of HLA alleles}
}
\value{
A vector with the same length as \code{allele}, with all deleted/changed
alleles converted to their new equivalent
}
\description{
Some HLA alleles were assigned in the past, but this assignment later turned
out to be wrong. For example, the allele was \emph{deleted} because it turned out
to be identical to an existing allele. Or, the allele was \emph{changed} by having
its sequence renamed (most commonly, a suffix is added).

\code{convert_deleted()} returns the new assignment of an HLA allele in the
\link{deleted_changed} lookup table. If the allele does not occur in the table
(i.e., it was never re-assigned), it simply returns the input allele. N.B. in
some cases, it was decided to remove the allele entirely, in which case \code{NA}
is returned.
}
\examples{
convert_deleted("B*08:06") # sequence identical to existing allele
convert_deleted("C*07:226") # sequence renamed (Q suffix added)
convert_deleted("C*07:295") # sequence never assigned (returns NA)
}
\seealso{
\code{\link[=convert_v2_to_v3]{convert_v2_to_v3()}} to lookup the v3 equivalent of v2 alleles
}
\keyword{internal}
