% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alleles.R
\name{count_alleles}
\alias{count_alleles}
\title{Count the number of alleles for each locus in an HLA typing string}
\usage{
count_alleles(
  typings,
  loci = c("A", "B", "C", "DPB1", "DQA1", "DQB1", "DRB1", "DRB.")
)
}
\arguments{
\item{typings}{A(n) (character vector of) HLA typing string(s).}

\item{loci}{A string or character vector with the loci you are interested in.
Only these alleles will be returned. Defaults to all. \code{DRB.} is used for
DRB3, DRB4, and DRB5.}
}
\value{
A (list of) named integer vector(s), with the loci as names, and
the number of found alleles per locus as values.
}
\description{
\code{count_alleles()} takes in a character vector or string of HLA typings, and
returns the number of alleles that the string contains for each locus.

This can be useful when validating whether a typing string contains a typing
for each locus. Also, this function can alert you when a typing string
contains more than two alleles for each locus, which can be intentional
(e.g. when the typing contains both the broad and the split) or a mistake.
}
\examples{
typing <- "A1 A2 B7 B8 Cw3 DQ5 DQ8 DR4 DR11 DR52 DR53"
count_alleles(typing, loci = "A")
count_alleles(typing)

# Also works with character vectors
typing <- c("A1 A2 B7 B8 Cw3", "Cw3 Cw7", NA)
count_alleles(typing, loci = c("A", "B", "C"))
}
