% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_db.R
\name{db_initialize}
\alias{db_initialize}
\title{Load IPD-IMGT/HLA database}
\usage{
db_initialize(data_dir, imgt_version = "Latest")
}
\arguments{
\item{data_dir}{Path to the folder where the SQLite database (~600MB) is or
should be stored. If no database exists on disk, it will be downloaded.}

\item{imgt_version}{Four-digit version number of the
\href{https://github.com/ANHIG/IMGTHLA/releases}{database release} that should
be used. Defaults to the most recent release ("Latest").}
}
\value{
A Python database connection object, which should be passed to other
functions that make use of the database.
}
\description{
Use the Python \href{https://github.com/nmdp-bioinformatics/}{\code{py-ard} package}
developed by the NMDP to initialize the
\href{https://www.ebi.ac.uk/ipd/imgt/hla/}{IPD-IMGT/HLA database}.
}
\examples{
\dontrun{
ard <- db_initialize(data_dir = "~/ipd_db/", imgt_version = "3510")
}
}
