% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deleted_changed}
\alias{deleted_changed}
\title{Table of changed/deleted HLA alleles}
\format{
\subsection{\code{deleted_changed}}{

A data frame with 275 rows and 2 columns:
\describe{
\item{allele_old}{allele name that was deleted/changed}
\item{allele_new}{current allele name}
}
}
}
\source{
\itemize{
\item \url{https://hla.alleles.org/alleles/deleted.html}
\item csv version of this data (used as the source)
\url{https://github.com/ANHIG/IMGTHLA/blob/Latest/Nomenclature_2009.txt}
}
}
\usage{
deleted_changed
}
\description{
The IPD-IMGT/HLA database contains a list of alleles in the database that
were once deleted (because, for example, they were later discovered to be
identical to an existing allele) or changed (for example, their sequence was
renamed). This conversion table lists all the old alleles and their new
equivalents.
}
\details{
The new names have been parsed from a description of the change contained in
the IPD-IMGT/HLA allele list. This makes it easier to programmatically
convert deleted/changed alleles without having to manually look up the new
designation.

N.B. The deleted alleles list is updated from time to time with new releases
of the IPD-IMGT/HLA database (though relatively infrequently; definitely
not every release). It's possible the conversion table in this package is
based on an older version.
}
\keyword{datasets}
