% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling.R
\name{etrl_lookup}
\alias{etrl_lookup}
\title{Get serological equivalents of an HLA-allele}
\usage{
etrl_lookup(allele)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A data frame with as many rows as there were elements in the input,
and four columns:
\enumerate{
\item The allele that was looked up (or \code{NA} if it didn't exist)
\item Its equivalent at the serological \emph{split} level (or \code{NA} if it doesn't
have one)
\item Its equivalent at the serological \emph{broad} level (or \code{NA} if it doesn't
have one)
\item Whether the allele has the public epitope (Bw4 or Bw6) (or \code{NA} if it
doesn't)
}
}
\description{
\code{etrl_lookup()} takes in a string or character vector of HLA alleles, and
returns their serological equivalents as defined in the ETRL HLA tables.
}
\details{
This function uses the EuroTransplant Reference Laboratory HLA (\link{etrl_hla})
tables to do the lookup. These tables define several alleles for each protein
(e.g. \code{HLA-A*01:01}, \code{HLA-A*01:02}), and their serological equivalents. All
others are grouped into an XX code (\code{HLA-A*01:XX}), which is also mapped to
a serological equivalent.
}
\section{Workings}{


All entered alleles will first be reduced to the two-field level (see
\code{\link[=reduce_to_nth_field]{reduce_to_nth_field()}}). If this reduced allele occurs in the lookup table,
the corresponding rows are returned. If not, it will be converted into an XX
code, and that row will be returned.
}

\section{Exceptions}{


If the allele is already in serological notation (e.g. \code{A1}), the lookup
fails and empty rows are returned (though e.g. \code{A1} is of course present
in the other columns of the lookup table)

If the allele has a suffix (e.g. \verb{HLA-C*01:37N}), it has no serological
equivalent, and hence will also return empty rows.
}

\examples{
allele_vec <- c(
  "B15", "B*15:79N", "B*15:YETY",
  "B*15:01:16", "B*15:02", "B*15:85"
)
etrl_lookup(allele_vec)
}
\seealso{
\link{etrl_hla}: the lookup table that's used by and returned in this
function
}
