% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mismatches.R
\name{get_mismatches}
\alias{get_mismatches}
\title{Determine mismatched antigens from typings}
\usage{
get_mismatches(donor, recipient)
}
\arguments{
\item{donor, recipient}{A pair of strings containing the donor and recipient
HLA typing, respectively. Each antigen must be separated by a space.}
}
\value{
A string with the mismatched antigens (separated by spaces).
}
\description{
\code{get_mismatches()} returns mismatched HLAs given a donor and recipient HLA
typing. That is, those antigens that occur in the \emph{donor} but \emph{not} the
recipient typing.
}
\examples{
donor <- "A1 A2"
recipient <- "A1 A2"
get_mismatches(donor, recipient)
#> ""
get_mismatches("A1 A2", "A3 A4")
#> "A1 A2"
get_mismatches("A1 A2 B5", "A3 A4 B5 B12")
#> "A1 A2"
}
