% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet_registry.R
\name{get_positive_eplets}
\alias{get_positive_eplets}
\title{Lookup and filter eplets unique to positive beads}
\usage{
get_positive_eplets(
  luminex_df,
  sample_col,
  alleles_col,
  assignment_col,
  eplet_df,
  pos_col = "eplets_pos"
)
}
\arguments{
\item{luminex_df}{Data frame with Luminex assay results.}

\item{sample_col}{Name of column in \code{luminex_df} containing a character
vector of sample IDs.}

\item{alleles_col}{Name of column in \code{luminex_df} containing a character
vector of bead specificities (i.e. the HLA coated on the bead).}

\item{assignment_col}{Name of column in \code{luminex_df} containing a logical
vector with the bead assignment (TRUE = positive, FALSE = negative).}

\item{eplet_df}{Data frame containing the Eplet Registry; from output of
\code{\link[=load_eplet_registry]{load_eplet_registry()}}.}

\item{pos_col}{Name of new column to contain positive-only eplets.}
}
\value{
Data frame with two columns, and one row for each eplet:
\enumerate{
\item Character vector of sample IDs as in \code{sample_col}
\item Character vector of eplet names.
}
}
\description{
\code{get_positive_eplets()} takes in results from a Luminex single antigen bead
assay, and extracts only the eplets that occur exclusively on positive- but
not negative beads for each sample. That is, the set of eplets on negative
beads is "subtracted" from the set of eplets on positive beads.
}
\examples{
\dontrun{
df_eplets <- load_eplet_registry()
luminex_df <- dplyr::tribble(
  ~sampleID, ~allele, ~positive,
  "001", "A*01:01", TRUE,
  "001", "A*02:01", FALSE,
  "002", "A*01:01", TRUE
)
get_positive_eplets(luminex_df, sampleID, allele, positive, df_eplets)
}
}
\seealso{
\code{\link[=lookup_eplets]{lookup_eplets()}} is used internally to retrieve the eplets on each
allele from the HLA Registry
}
