% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resolution.R
\name{get_resolution}
\alias{get_resolution}
\title{Determine the level of resolution of an HLA-allele}
\usage{
get_resolution(allele)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A string or character vector of the same length as \code{allele},
with \code{"low"}, \code{"intermediate"}, or \code{"high"} for each element.
}
\description{
\code{get_resolution()} takes in a string or character vector of HLA alleles, and
returns their resolution as either low, intermediate, or high.
}
\details{
Resolution is defined as follows:
\itemize{
\item \strong{Low}: a typing at the allele group level, such as
\itemize{
\item anything using the serological/antigen nomenclature, e.g \code{A2}
\item an XX code representing an allele group, e.g. \code{A*02:XX}
}
\item \strong{Intermediate}: any typing with ambiguities, such as
\itemize{
\item a list using the forward slash notation, e.g. \code{HLA-A*23:26/HLA-A*23:39}
\item a typing with multiple allele codes, e.g. \code{A*01:AABJE}
}
\item \strong{High}: any typing with more than 2 field codes, e.g.
\itemize{
\item \code{A*24:09} (minimum)
\item \code{HLA-A*24:02:01:02L} (maximum)
}
}
}
\examples{
get_resolution("A2") # low
get_resolution("A*01:AABJE") # intermediate
get_resolution("A*24:09") # high

# also works with character vectors, or in a data frame
allele_vec <- c("A2", "A*01:AABJE", "B*42:08")
get_resolution(allele_vec)

tidyr::tibble(alleles = allele_vec) |>
  dplyr::mutate(resolution = get_resolution(allele_vec))

}
