% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling.R
\name{get_split}
\alias{get_split}
\title{Retrieve split-level serological equivalent of an HLA-allele}
\usage{
get_split(allele)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A string or character vector of the same length as \code{allele},
with the corresponding split if it exists, or \code{NA} if none exists.
}
\description{
\code{get_split()} takes in a string or character vector of HLA alleles. The
corresponding split-level allele is looked up in \link{etrl_hla} and returned.
If no such allele exists, \code{NA} is returned instead.
}
\details{
This function uses the EuroTransplant Reference Laboratory HLA (\link{etrl_hla})
tables to do the lookup. These tables define several alleles for each protein
(e.g. \code{HLA-A*01:01}, \code{HLA-A*01:02}), and their serological equivalents. All
others are grouped into an XX code (\code{HLA-A*01:XX}), which is also mapped to
a serological equivalent.
\subsection{Workings}{

All entered alleles will first be reduced to the two-field level (see
\code{\link[=reduce_to_nth_field]{reduce_to_nth_field()}}). If this reduced allele occurs in the lookup table,
the corresponding rows are returned. If not, it will be converted into an XX
code, and that row will be returned.

If the allele is already at the serological broad- or split-level, the lookup
will be performed using those respective columns in \link{etrl_hla}.
}

\subsection{Exceptions}{

If the allele has a suffix (e.g. \verb{HLA-C*01:37N}), it has no serological
equivalent, and hence will also return nothing.
}
}
\examples{

get_split("A24") # is already a split, so returns itself ("A24")
get_split("A9") # is a broad, so returns `NA`

get_split("B*14:01") # returns corresponding split ("B64")
get_split("B*14:02") # returns corresponding split ("B65")
get_split("B*14:03") # no split is defined for this allele; returns `NA`

# Vectors also work:
get_split(c("A24", "B*14:01", "B*14:03"))
}
\seealso{
\itemize{
\item \link{get_broad}: for looking up the serological broad-level equivalent of an
allele
\item \link{get_serology}: for looking up the split, or broad if none exists
\item \link{get_public}: for looking up the public epitope of an allele
\item \link{etrl_hla}: the lookup table that's used in this function
}
}
