% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlapro-package.R
\docType{package}
\name{hlapro-package}
\alias{hlapro}
\alias{hlapro-package}
\title{hlapro: Tools to Work with HLA Typings and Single-Antigen Bead Assays}
\description{
Provides tools for Human Leukocyte Antigen (HLA) data: genotypes as well as (Luminex) single-antigen bead-based assays. Including: standardizing HLA typings to a common format (GL Strings), reducing and increasing HLA typing resolution, determining bead positivity and assessing donor-specific antibodies (DSAs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lcreteig/hlapro}
  \item \url{https://lcreteig.github.io/hlapro/}
  \item Report bugs at \url{https://github.com/lcreteig/hlapro/issues}
}

}
\author{
\strong{Maintainer}: Leon Reteig \email{leonreteig@gmail.com} (\href{https://orcid.org/0000-0002-5814-0992}{ORCID})

Other contributors:
\itemize{
  \item UMC Utrecht [copyright holder, funder]
}

}
\keyword{internal}
