% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet_registry.R
\name{load_eplet_registry}
\alias{load_eplet_registry}
\title{Load HLA Eplet Registry table}
\usage{
load_eplet_registry(
  folder_path = NULL,
  filename = "eplets.rds",
  print_version = TRUE,
  return_path = FALSE,
  delete = FALSE
)
}
\arguments{
\item{folder_path}{Character path to the folder where the previously
downloaded table is stored, or where you want a new version to be stored.
Defaults to the R user cache directory.}

\item{filename}{Character filename for the table.}

\item{print_version}{Logical indicating whether to print message with version
information when loading the table. Turn this off by setting to \code{FALSE}.}

\item{return_path}{Logical indicating whether to return the folder path
rather than the data frame.}

\item{delete}{Logical indicating whether to delete the cached Eplet Registry.}
}
\value{
Either a data frame containing the Eplet Registry (default) or a
character string giving the path to the folder where it is/will be stored
(if \code{return_path} = \code{TRUE}).
}
\description{
\code{load_eplet_registry()} returns a dataframe with the \href{https://www.epregistry.com.br}{HLA Eplet Registry table}, which maps HLA alleles to eplets. If
no local copy exists, the function will prompt you whether you want to scrape
the online registry to download a fresh copy.
}
\details{
Currently incorporates only the HLA tables (A/B/C, DRB, DQ, DP and
Interlocus), not MICA.
}
\examples{
\dontrun{
# (Down)load the eplet registry (to/)from the default path
df_eplets <- load_eplet_registry()
# Return the location of the default path
load_eplet_registry(return_path = TRUE)
# (Down)load the eplet registry (to/)from a custom path and a custom filename
df_eplets <- load_eplet_registry("~/eplet_registry", "eplet_table.rds")
# Delete the saved table
load_eplet_registry("~/eplet_registry", "eplet_table.rds", delete = TRUE)
}

}
\seealso{
\itemize{
\item \code{\link[=lookup_alleles]{lookup_alleles()}} for a helper function that takes in eplets and looks
up the HLA alleles they occur on in the registry
\item \code{\link[=lookup_eplets]{lookup_eplets()}} for a helper function that takes in alleles and looks
up which eplets occur on them in the registry
}
}
