% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet_registry.R
\name{lookup_eplets}
\alias{lookup_eplets}
\title{Lookup eplets corresponding to alleles in HLA Eplet Registry}
\usage{
lookup_eplets(eplet_df, alleles)
}
\arguments{
\item{eplet_df}{Data frame containing the Eplet Registry; from output of
\code{\link[=load_eplet_registry]{load_eplet_registry()}}.}

\item{alleles}{String or character vector of HLA alleles.}
}
\value{
Named list, where each element is a character vector of eplets for
each allele in the input.
}
\description{
\code{lookup_eplets()} takes in a set of HLA alleles, and retrieves the
corresponding eplets from the Eplet Registry table.
}
\examples{
\dontrun{
df_eplets <- load_eplet_registry()
lookup_eplets(df_eplets, "A*01:01")
# Also works for vectors:
lookup_eplets(df_eplets, c("A*01:01", "B*08:01"))
}
}
\seealso{
\code{\link[=lookup_alleles]{lookup_alleles()}} for the reverse operation
}
