% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_db.R
\name{mac_decode}
\alias{mac_decode}
\title{Decode a Multiple Allele Code (MAC) into an ambiguous HLA typing}
\usage{
mac_decode(ard, allele)
}
\arguments{
\item{ard}{The Python database connection object created by \code{\link[=db_initialize]{db_initialize()}}}

\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A string or character vector of the same length as \code{allele},
with the ambiguities.
}
\description{
\code{mac_decode()} is a thin wrapper around the \code{ard.expand_mac()} method of the
\href{https://github.com/nmdp-bioinformatics/py-ard}{\code{py-ard}} Python package.

See the \href{https://hml.nmdp.org/MacUI/}{NMDP website} for more on MACs.
}
\details{
Uses the \href{https://www.ebi.ac.uk/ipd/imgt/hla/}{IPD-IMGT/HLA database} as
initialized with \code{\link[=db_initialize]{db_initialize()}}. This function will throw an error if any
of the alleles in the input do not exist in the database. Use
\code{\link[=is_in_ipd_db]{is_in_ipd_db()}} to safely check if the allele(s) exist(s).
}
\examples{
\dontrun{
ard <- db_initialize(data_dir = path.expand("~/ipd_db"))
mac_decode(ard, "A*01:AB")
#> "A*01:01/A*01:02"
mac_decode(ard, "HLA-A*25:BYHR")
#> "HLA-A*25:01/HLA-A*26:01"
# Also works for vectors:
mac_decode(ard, c("A*01:AB", "A*25:BYHR"))
#> "A*01:01/A*01:02" "A*25:01/A*26:01"
}
}
\seealso{
\code{\link[=mac_lookup]{mac_lookup()}} for the reverse operation
}
