% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling.R
\name{reorder_alleles}
\alias{reorder_alleles}
\title{Swap a pair of HLA-alleles to match another}
\usage{
reorder_alleles(in_order, to_order)
}
\arguments{
\item{in_order, to_order}{A character vector of two HLA alleles of a certain
locus. The order of \code{to_order} will be reversed, to match \code{in_order}, if
necessary.}
}
\value{
A vector with the same alleles as \code{to_order}, possible reordered.
}
\description{
\code{reorder_alleles()} takes in two length-2 character vectors of HLA alleles,
and if they're not in the same order, reorders the 2nd to match the 1st.
This can be useful when you have multiple sources of HLA typings for one
individual that you're trying to match up.
}
\details{
If the vectors contain alleles in different nomenclatures / of different
resolutions, the order is determined based on their serological broad-level
equivalents.
}
\examples{
reorder_alleles(in_order = c("A1", "A2"), to_order = c("A2", "A1"))
# can accommodate missing/homozygous typings
reorder_alleles(in_order = c("A1", NA), to_order = c("A2", "A1"))
# still works if the typings are in a different format/resolution
reorder_alleles(in_order = c("A1", "A2"), to_order = c("A*02:01", "A*01:01"))
}
\keyword{internal}
