% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling.R
\name{strip_broad}
\alias{strip_broad}
\title{Strip broad in \code{split}(\code{broad}) notation}
\usage{
strip_broad(allele, check_result = FALSE)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}

\item{check_result}{If TRUE, returns input as is if the result of removing
the parentheses does not result in an existing split in \link{etrl_hla}. Sometimes
splits and broads are reversed (e.g. \code{A9(A24)}), in which case we don't want
to proceed.}
}
\value{
a vector of same length as allele
}
\description{
Removes parentheses and their contents, in e.g. \code{A24(A9)}. Because this
notation is not accepted by \code{\link[=validate_allele]{validate_allele()}}, and broads can always be
added back with \code{\link[=get_broad]{get_broad()}}.

TODO: should at some point be subsumed in a general \code{hla_clean()} type
function
}
\examples{
strip_broad("A24(A9)")
}
\keyword{internal}
