% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hla.R
\name{strip_redundant}
\alias{strip_redundant}
\title{Strip redundant alleles from typing string if higher resolution is available}
\usage{
strip_redundant(typing)
}
\arguments{
\item{typing}{A string containing the HLA allele or the full space-separated
HLA typing.}
}
\value{
A string without the redundant alleles in the input
}
\description{
Sometimes a typing will contain multiple "versions" of an allele at different
levels of resolution. Most commonly this is both the split and the broad,
e.g. \code{A24(9)} or \verb{A10 A25}. The latter can cause a typing to contain more
than 2 alleles for a given locus, which cannot be handled with
\code{\link[=extract_alleles_str]{extract_alleles_str()}}; the former is not accepted by \code{\link[=validate_allele]{validate_allele()}}.
Besides, the broads are redundant in this case, and can always be added back
with \code{\link[=get_broad]{get_broad()}}. Sometimes even three "versions" are included: the split,
the broad, and the two-field allele, e.g. \verb{DR5 DR11 DRB1*11:03}.
}
\examples{
strip_redundant("A24(9)")
strip_redundant("A9(24)") # also works when the split is in parentheses
strip_redundant("A24(9) A10 A25") # removes both A9 and A10
strip_redundant("DR5 DR11 DRB1*11:03") # removes both the split and the broad
# also works on character vectors
strip_redundant(c("A24(9)", "A25(10)"))
}
\keyword{internal}
