% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v2_to_v3}
\alias{v2_to_v3}
\title{v2 to v3 HLA nomenclature conversion table}
\format{
\subsection{\code{v2_to_v3}}{

A data frame with 5,151 rows and 2 columns:
\describe{
\item{v2}{v2 version of the allele}
\item{v3}{v3 version of the allele}
}
}
}
\source{
\itemize{
\item Most pre-2010 alleles:
\url{https://github.com/ANHIG/IMGTHLA/blob/Latest/Nomenclature_2009.txt}
\item Additional mappings for deleted alleles:
\url{https://github.com/ANHIG/IMGTHLA/blob/Latest/Deleted_alleles.txt}
\item Allele-specific & DPB1-specific multiple-allele codes (see 4. and 5. under
"Allele Code Update to Version 3 WHO Nomenclature")
N.B. Unfortunately this link no longer works
\url{https://bioinformatics.bethematchclinical.org/hla-resources/allele-codes/}
}
}
\usage{
v2_to_v3
}
\description{
In April 2010, the HLA nomenclature officially changed from "v2" to "v3",
which mandates the use of colons ("\code{:}") to delimit fields. For example:
\code{A*01010101} (v2) is since written as \code{A*01:01:01:01}. Note though that not
all conversions are this straightforward, there are a number of exceptions.

This conversion table tries to collect all standard v2 --> v3 conversions
from different data sources (see below).
}
\keyword{datasets}
