% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alleles.R
\name{vec_to_gl}
\alias{vec_to_gl}
\title{Convert a vector of HLA alleles to a GL String}
\usage{
vec_to_gl(allele_list, namespace = "hla", version_or_date = NULL)
}
\arguments{
\item{allele_list}{a character vector of HLA alleles.}

\item{namespace}{\href{https://glstring.org/syntax-1.1.html#namespace}{Specification} of the
HLA nomenclature system. Defaults to \code{"hla"}.}

\item{version_or_date}{\href{https://glstring.org/syntax-1.1.html#versionordate}{Specification} of the
version of the \code{namespace} used, or the date the GL String was constructed.
If not specified, uses today's date as the default.}
}
\value{
A valid GL String.
}
\description{
\code{vec_to_gl()} constructs a \href{https://glstring.org}{GL String} from a character
vector of HLAs, as well as some metadata (namespace, version/data).
}
\examples{
vec_to_gl(c("A*02:302", "A*23:26/A*23:39", "B*44:02:13", "B*49:08"))
vec_to_gl(
  allele_list = c("HLA-A*02:69", "HLA-A*23:30"),
  namespace = "hla", version_or_date = "2018-06"
)
}
\seealso{
\itemize{
\item \code{\link[=gl_to_vec]{gl_to_vec()}} for the opposite operation
\item \code{\link[=df_to_gl]{df_to_gl()}} for converting a dataframe with one column per locus/allele to
a GL String
}
}
\keyword{internal}
