% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_db.R
\name{db_get_version}
\alias{db_get_version}
\title{Print version of initialized IPD-IMGT/HLA database}
\usage{
db_get_version(ard)
}
\arguments{
\item{ard}{The Python database connection object created by \code{\link[=db_initialize]{db_initialize()}}}
}
\description{
Thin wrapper around the \code{ard.db_get_version()} method of the
\href{https://github.com/nmdp-bioinformatics/py-ard}{\code{py-ard}} Python package.
}
\examples{
\dontrun{
ard <- db_initialize(data_dir = "~/ipd_db/")
db_get_version(ard)
}
}
\seealso{
\code{\link[=db_initialize]{db_initialize()}}
}
