% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling.R
\name{etrl_lookup}
\alias{etrl_lookup}
\title{Get rows with serological equivalents of an HLA allele from ETRL HLA table}
\usage{
etrl_lookup(allele)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A data frame with as many rows as there were elements in the input,
and four columns:
\enumerate{
\item The allele that was looked up (or \code{NA} if it didn't exist)
\item Its equivalent at the serological \emph{split} level (or \code{NA} if it doesn't
have one)
\item Its equivalent at the serological \emph{broad} level (or \code{NA} if it doesn't
have one)
\item Whether the allele has the public epitope (Bw4 or Bw6) (or \code{NA} if it
doesn't)
}
}
\description{
\code{etrl_lookup()} takes in a string or character vector of HLA alleles, and
returns their serological equivalents as defined in the ETRL HLA tables.
}
\examples{
allele_vec <- c(
  "B15", "B*15:79N", "B*15:YETY",
  "B*15:01:16", "B*15:02", "B*15:85"
)
etrl_lookup(allele_vec)
}
\seealso{
\link{etrl_hla}: the lookup table that's used by and returned in this
function
}
\keyword{internal}
