% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling.R
\name{get_broad}
\alias{get_broad}
\title{Retrieve broad-level equivalent of a split-level HLA-allele}
\usage{
get_broad(allele)
}
\arguments{
\item{allele}{A string or character vector with (an) HLA allele(s).}
}
\value{
A string or character vector of the same length as \code{allele},
with the corresponding broad if it exists, or \code{NA} if none exists.
}
\description{
\code{get_broad()} takes in a string or character vector of HLA alleles. The
corresponding broad-level allele is looked up in \link{etrl_hla} and returned.
If no such allele exists, \code{NA} is returned instead.
}
\examples{
get_broad("A24") # returns corresponding broad ("A9")
get_broad("A9") # is already a broad, returns itself ("A9")

# these alleles in modern nomenclature also all return "A9"
get_broad("A*24")
get_broad("A*24:XX")
get_broad("A*24:02:01:102")

# Vectors also work:
get_broad(c("A24", "A23", "A1"))
}
\seealso{
\itemize{
\item \link{get_public}: for looking up the public epitope of an allele
\item \link{etrl_hla}: the lookup table that's used in this function
\item \link{etrl_lookup}: for getting serological equivalents of other alleles
}
}
