% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alleles.R
\name{gl_to_vec}
\alias{gl_to_vec}
\title{Split a GL String into metadata and a vector of alleles}
\usage{
gl_to_vec(glstring)
}
\arguments{
\item{glstring}{Any valid GL String.}
}
\value{
A named list of length three containing the:
\enumerate{
\item \code{namespace} (e.g. \code{"hla"})
\item \code{version_or_date} (e.g. \code{"3.29.0"} or \code{"2023-05-27"})
\item \code{allele_list} (e.g. \code{c("HLA-A*01:01", "B*07:08")})
}
}
\description{
\code{gl_to_vec()} takes in a \href{https://glstring.org}{GL String} and converts it to
a list containing the GL String metadata (namespace, version/data) and a
character vector of the individual HLA alleles.
}
\examples{
# return list that includes the metadata
gl_to_vec("hla#2018-06#HLA-A*02:69+HLA-A*23:30")
# access the allele list directly to keep just the vector of HLAs
gl_to_vec("hla#2018-06#HLA-A*02:69+HLA-A*23:30")$allele_list
}
\seealso{
\itemize{
\item \code{\link[=vec_to_gl]{vec_to_gl()}} for the opposite operation
\item \code{\link[=gl_to_df]{gl_to_df()}} for converting a vector of GL Strings to a dataframe with one
column per locus/allele
}
}
\keyword{internal}
